pyproject.toml
[project]
name = "infovail-iq"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.84.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "tenacity>=9.1.4",
]
